+++ probe-rs/src/debug/debug_info.rs
@@ -2,9 +2,10 @@ use super::{
     function_die::FunctionDie, get_sequential_key, unit_info::UnitInfo, unit_info::UnitIter,
     variable::*, DebugError, DebugRegisters, SourceLocation, StackFrame, VariableCache,
 };
+use crate::core::UnwindRule;
 use crate::{
     core::Core,
-    core::{RegisterRole, RegisterValue},
+    core::{ExceptionInfo, ExceptionInterface, RegisterRole, RegisterValue},
     debug::{registers, source_statement::SourceStatements},
     MemoryInterface,
 };
@@ -497,15 +498,22 @@ impl DebugInfo {
         &self,
         core: &mut Core<'_>,
         address: u64,
+        exception_info: &Option<ExceptionInfo>,
         unwind_registers: &registers::DebugRegisters,
     ) -> Result<Vec<StackFrame>, DebugError> {
         let mut units = self.get_units();
 
-        let unknown_function = format!(
-            "<unknown function @ {:#0width$x}>",
-            address,
-            width = (unwind_registers.get_address_size_bytes() * 2 + 2)
-        );
+        let unknown_function = if let Some(exception_info) = exception_info {
+            exception_info.description.to_string()
+        } else {
+            // When reporting the address, we format it as a hex string, with the width matching
+            // the configured size of the datatype used in the `RegisterValue` address.
+            format!(
+                "<unknown function @ {:#0width$x}>",
+                address,
+                width = (unwind_registers.get_address_size_bytes() * 2 + 2)
+            )
+        };
         let stack_frame_registers = unwind_registers.clone();
 
         let mut frames = Vec::new();
@@ -692,7 +700,6 @@ impl DebugInfo {
     /// The unwind loop will continue until we meet one of the following conditions:
     /// - We can no longer unwind a valid PC value to be used for the next frame.
     /// - We encounter a LR register value of 0x0 or 0xFFFFFFFF(Arm 'Reset' value for that register).
-    /// - TODO: Catch the situation where the PC value indicates a hard-fault or other non-recoverable exception
     /// - We can not intelligently calculate a valid LR register value from the other registers, or the gimli::RegisterRule result is a value of 0x0. Note: [DWARF](https://dwarfstd.org) 6.4.4 - CIE defines the return register address used in the `gimli::RegisterRule` tables for unwind operations. Theoretically, if we encounter a function that has `Undefined` `gimli::RegisterRule` for the return register address, it means we have reached the bottom of the stack OR the function is a 'no return' type of function. I have found actual examples (e.g. local functions) where we get `Undefined` for register rule when we cannot apply this logic. Example 1: local functions in main.rs will have LR rule as `Undefined`. Example 2: main()-> ! that is called from a trampoline will have a valid LR rule.
     /// - Similarly, certain error conditions encountered in `StackFrameIterator` will also break out of the unwind loop.
     /// Note: In addition to populating the `StackFrame`s, this function will also populate the `DebugInfo::VariableCache` with `Variable`s for available Registers as well as static and function variables.
@@ -716,18 +723,46 @@ impl DebugInfo {
             .get_program_counter()
             .and_then(|pc| pc.value)
         {
+            // PART 0: The first step is to determine the exception context for the current PC.
+            // - If we are at an exception hanlder frame, we need to overwrite the unwind registers with the exception context.
+            // - If for some reason we cannot determine the exception context, we silently continue with the rest of the unwind.
+            // At worst, the unwind will be able to unwind the stack to the frame of the most recent exception handler.
+            let exception_info = match core.exception_details(&unwind_registers) {
+                Ok(Some(exception_info)) => {
+                    tracing::trace!(
+                        "UNWIND: Found exception context: {}",
+                        exception_info.description
+                    );
+                    Some(exception_info)
+                }
+                Ok(None) => {
+                    tracing::trace!(
+                        "UNWIND: No exception context found. Stack unwind will continue."
+                    );
+                    None
+                }
+                Err(e) => {
+                    tracing::warn!("UNWIND: Error while checking for exception context. Unwinding will stop at the first exception handler. : {}", e);
+                    None
+                }
+            };
+
             // PART 1: Construct the `StackFrame` for the current pc.
             let frame_pc = frame_pc_register_value
                 .try_into()
-                .map_err(|error| crate::Error::Other(anyhow::anyhow!("Cannot convert register value for program counter to a 64-bit integeer value: {:?}", error)))?;
+                .map_err(|error| crate::Error::Register(format!("Cannot convert register value for program counter to a 64-bit integeer value: {:?}", error)))?;
             tracing::trace!(
                 "UNWIND: Will generate `StackFrame` for function at address (PC) {}",
                 frame_pc,
             );
 
-            //
             // PART 1-a: Prepare the `StackFrame` that holds the current frame information.
-            let return_frame = match self.get_stackframe_info(core, frame_pc, &unwind_registers) {
+            let return_frame = match self.get_stackframe_info(
+                core,
+                frame_pc,
+                &exception_info,
+                &unwind_registers,
+            ) {
                 Ok(mut cached_stack_frames) => {
                     while cached_stack_frames.len() > 1 {
                         // If we encountered INLINED functions (all `StackFrames`s in this Vec, except for the last one, which is the containing NON-INLINED function), these are simply added to the list of stack_frames we return.
@@ -759,13 +794,40 @@ impl DebugInfo {
             };
 
             // Part 1-b: Check LR values to determine if we can continue unwinding.
-            // TODO: ARM has special ranges of LR addresses to indicate fault conditions. We should check those also.
             if let Some(check_return_address) = unwind_registers.get_return_address() {
                 if check_return_address.is_max_value() || check_return_address.is_zero() {
                     // When we encounter the starting (after reset) return address, we've reached the bottom of the stack, so no more unwinding after this.
                     stack_frames.push(return_frame);
                     tracing::trace!("UNWIND: Stack unwind complete - Reached the 'Reset' value of the LR register.");
                     break;
+                } else {
+                    // Part 1-c: If the target current frame is an exception handler, we need to update the `unwind_registers` to match the frame that invoked the exception handler.
+                    if let Some(exception_info) = exception_info {
+                        tracing::trace!(
+                            "UNWIND: Stack unwind reached an exception handler {}",
+                            exception_info.description
+                        );
+                        if let Some(exception_info) = core.exception_details(&unwind_registers)? {
+                            tracing::trace!(
+                                "UNWIND: Found exception context: {}",
+                                exception_info.description
+                            );
+                            // If we are at an exception hanlder frame, we need to overwrite the unwind registers.
+                            // This will allow us to continue unwinding from the exception handler frame.
+                            unwind_registers = exception_info.calling_frame_registers;
+
+                            // Now that we've optionally updated the `unwind_registers` to match the exception handler, we can continue.
+                            stack_frames.push(return_frame);
+                            tracing::trace!(
+                                    "UNWIND: Stack unwind will attempt to unwind the frame that invoked {}.", exception_info.description
+                                );
+                            continue;
+                        } else {
+                            tracing::trace!(
+                                "UNWIND: No exception context found. Stack unwind will continue."
+                            )
+                        }
+                    }
                 }
             } else {
                 // If the debug info rules result in a None return address, we cannot continue unwinding.
@@ -1244,34 +1306,35 @@ fn unwind_register(
                     // NOTE: PC = Value of the unwound LR, i.e. the first instruction after the one that called this function.
                     register_rule_string = "PC=(unwound LR) (dwarf Undefined)".to_string();
                     unwound_return_address.and_then(|return_address| {
-                        if return_address.is_max_value() || return_address.is_zero() {
-                            tracing::warn!("No reliable return address is available, so we cannot determine the program counter to unwind the previous frame.");
-                            None
-                        } else {
-                            match return_address {
-                                RegisterValue::U32(return_address) => {
-                                    if matches!(core.instruction_set(), Ok(InstructionSet::Thumb2)) {
-                                        // NOTE: [ARMv7-M Architecture Reference Manual](https://developer.arm.com/documentation/ddi0403/ee), Section A5.1.2: We have to clear the last bit to ensure the PC is half-word aligned. (on ARM architecture, when in Thumb state for certain instruction types will set the LSB to 1)
-                                        register_rule_string = "PC=(unwound LR & !0b1) (dwarf Undefined)".to_string();
-                                        Some(RegisterValue::U32(return_address  & !0b1))
-                                    } else{
-                                        Some(RegisterValue::U32(return_address))
-                                    }
-                                }
-                                RegisterValue::U64(return_address) => {
-                                    Some(RegisterValue::U64(return_address))
-                                },
-                                RegisterValue::U128(_) => {
-                                    tracing::warn!("128 bit address space not supported");
-                                    None
-                                }
-                            }
-                        }
+                        unwind_program_counter_register(
+                            return_address,
+                            core,
+                            &mut register_rule_string,
+                        )
                     })
                 }
-                _ => {
-                    // This will result in the register value being cleared for the previous frame.
-                    None
+                other_register => {
+                    // If the the register rule was not specified, then we either carry the previous value forward,
+                    // or we clear the register value, depending on the architecture and register type.
+                    match other_register.core_register.unwind_rule {
+                        UnwindRule::Preserve => {
+                            register_rule_string = "Preserve".to_string();
+                            callee_frame_registers
+                                .get_register(other_register.core_register.id)
+                                .and_then(|reg| reg.value)
+                        }
+                        UnwindRule::Clear => {
+                            register_rule_string = "Clear".to_string();
+                            None
+                        }
+                        UnwindRule::SpecialRule => {
+                            // When no DWARF rules are available, and it is not a special register like PC, SP, FP, etc.,
+                            // we will preserve the value. It is possible it might have its value set later if
+                            // exception frame information is available.
+                            register_rule_string = "Clear (no unwind rules specified)".to_string();
+                            None
+                        }
+                    }
                 }
             }
         }
@@ -1352,6 +1415,35 @@ fn unwind_register(
         register_rule_string,
     );
     ControlFlow::Continue(())
+}
+
+/// Helper function to determine the program counter value for the previous frame.
+fn unwind_program_counter_register(
+    return_address: RegisterValue,
+    core: &mut Core<'_>,
+    register_rule_string: &mut String,
+) -> Option<RegisterValue> {
+    if return_address.is_max_value() || return_address.is_zero() {
+        tracing::warn!("No reliable return address is available, so we cannot determine the program counter to unwind the previous frame.");
+        None
+    } else {
+        match return_address {
+            RegisterValue::U32(return_address) => {
+                if matches!(core.instruction_set(), Ok(InstructionSet::Thumb2)) {
+                    // NOTE: [ARMv7-M Architecture Reference Manual](https://developer.arm.com/documentation/ddi0403/ee), Section A5.1.2: We have to clear the last bit to ensure the PC is half-word aligned. (on ARM architecture, when in Thumb state for certain instruction types will set the LSB to 1)
+                    *register_rule_string = "PC=(unwound LR & !0b1) (dwarf Undefined)".to_string();
+                    Some(RegisterValue::U32(return_address & !0b1))
+                } else {
+                    Some(RegisterValue::U32(return_address))
+                }
+            }
+            RegisterValue::U64(return_address) => Some(RegisterValue::U64(return_address)),
+            RegisterValue::U128(_) => {
+                tracing::warn!("128 bit address space not supported");
+                None
+            }
+        }
+    }
 }
 
 /// Helper function to handle adding a signed offset to a u64 address.